Extract methods from UIActionEngine to PackageSourceMappingUtility
