+++ src/NuGet.Clients/NuGet.PackageManagement.UI/Actions/UIActionEngine.cs
@@ -471,36 +471,7 @@ namespace NuGet.PackageManagement.UI
 
                     if (!cancellationToken.IsCancellationRequested)
                     {
-                        if (userAction?.SourceMappingSourceName != null && addedPackages != null)
-                        {
-                            Dictionary<string, IReadOnlyList<string>> patternsReadOnly = existingPackageSourceMappingSourceItems
-                                .ToDictionary(pair => pair.Key, pair => (IReadOnlyList<string>)(pair.Patterns.Select(p => p.Pattern).ToList()));
-
-                            PackageSourceMapping packageSourceMapping = new(patternsReadOnly);
-
-                            // Expand all patterns/globs so we can later check if this package ID was already mapped.
-                            var addedPackageIds = addedPackages.Select(action => action.Item1);
-                            var addedPackageIdsWithoutExistingMappings = new List<string>();
-                            foreach (string? addedPackageId in addedPackageIds)
-                            {
-                                var configuredSource = packageSourceMapping.GetConfiguredPackageSources(addedPackageId);
-                                if (configuredSource == null || configuredSource.Count == 0)
-                                {
-                                    addedPackageIdsWithoutExistingMappings.Add(addedPackageId);
-                                }
-                            }
-
-                            // Get all newly added package IDs that were not previously Source Mapped.
-                            // Always include the Package ID being installed since it takes precedence over any globbing.
-                            string[] packageIdsNeedingNewSourceMappings = addedPackageIdsWithoutExistingMappings
-                               .Union(new string[] { userAction.PackageId })
-                               .ToArray();
-
-                            CreateAndSavePackageSourceMappings(
-                                sourceName: userAction.SourceMappingSourceName,
-                                newPackageIdsToSourceMap: packageIdsNeedingNewSourceMappings,
-                                sourceMappingProvider);
-                        }
+                        PackageSourceMappingUtility.ConfigureNewPackageSourceMapping(userAction, addedPackages, sourceMappingProvider, existingPackageSourceMappingSourceItems);
 
                         await projectManagerService.ExecuteActionsAsync(
                             actions,
@@ -932,43 +903,6 @@ namespace NuGet.PackageManagement.UI
             }
 
             return results;
-        }
-
-        private static void CreateAndSavePackageSourceMappings(string sourceName, string[] newPackageIdsToSourceMap, PackageSourceMappingProvider mappingProvider)
-        {
-            if (string.IsNullOrWhiteSpace(sourceName) || newPackageIdsToSourceMap is null || newPackageIdsToSourceMap.Length == 0)
-            {
-                return;
-            }
-
-            IEnumerable<PackagePatternItem> newPackagePatternItems = newPackageIdsToSourceMap.Select(packageId => new PackagePatternItem(packageId));
-
-            IReadOnlyList<PackageSourceMappingSourceItem> existingPackageSourceMappingItems = mappingProvider.GetPackageSourceMappingItems();
-            List<PackageSourceMappingSourceItem> newAndExistingPackageSourceMappingItems = new(existingPackageSourceMappingItems);
-
-            PackageSourceMappingSourceItem existingPackageSourceMappingItemForSource =
-                        existingPackageSourceMappingItems
-                        .Where(mappingItem => mappingItem.Key == sourceName)
-                        .FirstOrDefault();
-
-            // Source is being mapped for the first time.
-            if (existingPackageSourceMappingItemForSource is null)
-            {
-                existingPackageSourceMappingItemForSource = new(sourceName, newPackagePatternItems);
-                newAndExistingPackageSourceMappingItems.Add(existingPackageSourceMappingItemForSource);
-            }
-            else // Source already had an existing mapping.
-            {
-                foreach (var newPattern in newPackagePatternItems)
-                {
-                    if (!existingPackageSourceMappingItemForSource.Patterns.Contains(newPattern))
-                    {
-                        existingPackageSourceMappingItemForSource.Patterns.Add(newPattern);
-                    }
-                }
-            }
-
-            mappingProvider.SavePackageSourceMappings(newAndExistingPackageSourceMappingItems);
         }
 
         // Non-private only to facilitate testing.
